requirements.txt
accelerate==1.12.0
aiofiles==24.1.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
audioread==3.1.0
av==16.1.0
brotli==1.2.0
certifi==2026.1.4
cffi==2.0.0
cfgv==3.5.0
charset-normalizer==3.4.4
click==8.3.1
coloredlogs==15.0.1
contourpy==1.3.2
cryptography==46.0.4
ctranslate2==4.7.1
cycler==0.12.1
decorator==4.4.2
distlib==0.4.0
dotenv==0.9.9
einops==0.8.2
exceptiongroup==1.3.1
fastapi==0.129.0
faster-whisper==1.2.1
ffmpy==1.0.0
filelock==3.20.3
flatbuffers==25.12.19
fonttools==4.61.1
fsspec==2026.2.0
google-auth==2.48.0
google-genai==1.47.0
gradio==6.5.1
gradio_client==2.0.3
groovy==0.1.2
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
huggingface_hub==0.36.2
humanfriendly==10.0
identify==2.6.16
idna==3.11
ImageIO==2.37.2
imageio-ffmpeg==0.6.0
iniconfig==2.3.0
Jinja2==3.1.6
joblib==1.5.3
json_repair==0.57.1
kiwisolver==1.4.9
lazy_loader==0.4
librosa==0.11.0
librt==0.7.8
llvmlite==0.46.0
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mdurl==0.1.2
more-itertools==10.8.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.2
mypy_extensions==1.1.0
networkx==3.4.2
nodeenv==1.10.0
numba==0.63.1
numpy==2.0.2
onnxruntime==1.23.2
openai-whisper==20250625
orjson==3.11.7
packaging==26.0
pandas==2.3.3
pathspec==1.0.4
pdf2image==1.17.0
pillow==11.3.0
platformdirs==4.5.1
pluggy==1.6.0
pooch==1.9.0
pre_commit==4.5.1
proglog==0.1.12
protobuf==6.33.5
psutil==7.2.2
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.3.2
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.22
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.3
qwen-tts==0.1.1
regex==2026.1.15
requests==2.32.5
rich==14.3.2
rsa==4.9.1
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.15.3
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
soundfile==0.13.1
sox==1.5.0
soxr==1.0.0
starlette==0.52.1
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
tomli==2.4.0
tomlkit==0.13.3
torch==2.10.0
torchaudio==2.10.0
torchvision==0.25.0
tqdm==4.67.2
transformers==4.57.3
typer==0.23.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.32.0
uvloop==0.22.1
virtualenv==20.36.1
watchfiles==1.1.1
websockets==15.0.1
xlsxwriter==3.2.9
